backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
requests
jinja2
xlsxwriter
pytest
